PythonProject/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = "'Playwright tests'"
authors = ["Alexei Dobronravov <dobrick2003@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
playwright = "1.44"
pytest = "8.3"
pytest-playwright = "0.5"
allure-pytest = "^2.15.3"
assertpy = "^1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

